+++ R/knitr_utils.R
@@ -95,7 +95,6 @@ knitr_opts_current <- function(x, default = FALSE){
   x
 }
 
-
 #' @export
 #' @title Table options in a 'knitr' context
 #' @description Get options for table rendering in a 'knitr'
@@ -104,8 +103,7 @@ knitr_opts_current <- function(x, default = FALSE){
 #' table options are documented when 'knitr' is used.
 #'
 #' The function is a utility to facilitate the retrieval of table
-#' options supported by the 'flextable', 'officedown' and of
-#' course 'officer' packages.
+#' options supported by the 'flextable', 'officedown' and 'officer' packages.
 #'
 #' These options should be set with `knitr::opts_chunk$set()`.
 #' The names and expected values are listed in the following sections.
@@ -132,11 +130,13 @@ knitr_opts_current <- function(x, default = FALSE){
 #'
 #' @section knitr chunk options for Word tables:
 #'
-#' | **label**                                                     |    **name**  | **value**  |
-#' |:--------------------------------------------------------------|:------------:|:----------:|
-#' | the Word stylename to use for tables                          | tab.style    |    NULL    |
-#' | autofit' or 'fixed' algorithm.                                | tab.layout   |  "autofit" |
-#' | value of the preferred width of the table in percent (base 1).| tab.width    |      1     |
+#' | **label**                                                     |       **name**      | **value**  |
+#' |:--------------------------------------------------------------|:-------------------:|:----------:|
+#' | the Word stylename to use for tables                          | tab.style           |    NULL    |
+#' | autofit' or 'fixed' algorithm.                                | tab.layout          |  "autofit" |
+#' | value of the preferred width of the table in percent (base 1).| tab.width           |      1     |
+#' | Alternative title text                                        | tab.alt.title       |    NULL    |
+#' | Alternative description text                                  | tab.alt.description |    NULL    |
 #'
 #' @section knitr chunk options for data.frame with officedown:
 #'
@@ -149,7 +149,9 @@ knitr_opts_current <- function(x, default = FALSE){
 #' | don't display odd and even rows                               | no_hband        |    TRUE    |
 #' | don't display odd and even columns                            | no_vband        |    TRUE    |
 #'
-#' @return a list with following elements:
+#' @return a list
+#'
+#' @section returned elements:
 #'
 #' * cap.style (default: NULL)
 #' * cap.pre (default: "Table ")
@@ -159,6 +161,8 @@ knitr_opts_current <- function(x, default = FALSE){
 #' * cap.fp_text (default: `fp_text_lite(bold = TRUE)`)
 #' * id (default: NULL)
 #' * cap (default: NULL)
+#' * alt.title (default: NULL)
+#' * alt.description (default: NULL)
 #' * topcaption (default: TRUE)
 #' * style (default: NULL)
 #' * tab.lp (default: "tab:")
@@ -173,18 +177,67 @@ knitr_opts_current <- function(x, default = FALSE){
 #' @family functions for officer extensions
 #' @keywords internal
 opts_current_table <- function() {
+
+  is_quarto <- isTRUE(knitr::opts_knit$get("quarto.version") > numeric_version("0"))
+  is_bookdown <- isTRUE(knitr::opts_knit$get('bookdown.internal.label'))
+
+  if (is_bookdown) {
+    tab.lp <- knitr_opts_current("tab.lp", default = "tab:")
+  } else if (is_quarto) {
+    tab.lp <- "tab:"
+  } else {
+    tab.lp <- knitr_opts_current("tab.lp", default = NULL)
+  }
+  if (is_quarto) {
+    tab.topcaption <- knitr_opts_current("tbl-cap-location", default = "top") %in% "top"
+  } else {
+    tab.topcaption <- knitr_opts_current("tab.topcaption", default = TRUE)
+  }
+
+  if (is_quarto) {
+    tab.cap.pre <- knitr_opts_current("tbl-title", default = "Table")
+    tab.cap.sep <- knitr_opts_current("title-delim", default = ":")
+  } else {
+    tab.cap.pre <- knitr_opts_current("tab.cap.pre", default = "Table")
+    tab.cap.sep <- knitr_opts_current("tab.cap.sep", default = ":")
+  }
+  tab.cap.pre <- paste0(tab.cap.pre, " ")
+  tab.cap.sep <- paste0(tab.cap.sep, " ")
+
+  if (is_quarto) {
+    tab.cap <- knitr_opts_current("tbl-cap", default = NULL)
+  } else {
+    tab.cap <- knitr_opts_current("tab.cap", default = NULL)
+  }
+
+  if (is_quarto) {
+    tab.id <- knitr_opts_current("label", default = NULL)
+    if(!is.null(tab.id)) {
+      if (grepl("^unnamed\\-chunk", tab.id)) {
+        tab.id <- NULL
+      }
+    }
+  } else if (is_bookdown) {
+    tab.id <- knitr_opts_current("label", default = NULL)
+    tab.id <- knitr_opts_current("tab.id", default = tab.id)
+  } else {
+    tab.id <- knitr_opts_current("tab.id", default = NULL)
+  }
+
+  if (is_quarto) {
+    if(is.null(tab.cap) && !is.null(tab.id)) {
+      stop("if a label (", tab.id, ") is defined, chunk option `tbl-cap` should also be defined.")
+    }
+  }
+
   tab.cap.style <- knitr_opts_current("tab.cap.style", default = NULL)
-  tab.cap.pre <- knitr_opts_current("tab.cap.pre", default = "Table ")
-  tab.cap.sep <- knitr_opts_current("tab.cap.sep", default = ": ")
+  tab.alt.title <- knitr_opts_current("tab.alt.title", default = NULL)
+  tab.alt.description <- knitr_opts_current("tab.alt.description", default = NULL)
 
   tab.cap.tnd <- knitr_opts_current("tab.cap.tnd", default = 0)
   tab.cap.tns <- knitr_opts_current("tab.cap.tns", default = "-")
   tab.cap.fp_text <- knitr_opts_current("tab.cap.fp_text", default = fp_text_lite(bold = TRUE))
 
-  tab.cap <- knitr_opts_current("tab.cap", default = NULL)
-  tab.topcaption <- knitr_opts_current("tab.topcaption", default = TRUE)
-  tab.id <- knitr_opts_current("tab.id", default = NULL)
-  tab.lp <- knitr_opts_current("tab.lp", default = "tab:")
   tab.style <- knitr_opts_current("tab.style", default = NULL)
   tab.layout <- knitr_opts_current("tab.layout", default = "autofit")
   tab.width <- knitr_opts_current("tab.width", default = 1)
@@ -204,6 +257,8 @@ opts_current_table <- function() {
     cap.tns = tab.cap.tns,
     cap.fp_text = tab.cap.fp_text,
     id = tab.id,
+    alt.title = tab.alt.title,
+    alt.description = tab.alt.description,
     topcaption = tab.topcaption,
     cap = tab.cap,
     style = tab.style,